app/requirements.txt
asgiref==3.5.2
Django==3.2.6
django-filter==22.1
djangorestframework
flake8==5.0.4
gunicorn==20.1.0
Markdown==3.4.1
mccabe==0.7.0
psycopg2-binary==2.9.1
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pytz==2022.5
PyYAML==6.0
sqlparse==0.4.3
pillow